A small patch to prevent possible armor dupe
This wouldn't occur normally, but with other mods which, for example, add damage from another DamageSource to player's punches, shift-attacking a dummy causes TargetDummyEntity#hurt (and, in turn, TargetDummyEntity#dismantle) to be called twice, duping the items equipped.
Sorry for not testing it myself and leaving it to you ;_; I only have 1.16.5 Forge mdk and my hard drive space is tight. Though I believe this patch should do the trick.